refactor: LadderFactory에 height에 관한 기능을 Height 객체로 옮긴다.
변경 내용:
String height를 int 로 바꾸는 기능을 옮긴다.
int height의 범위를 검증하는 기능을 옮긴다.
Ladder 생성 시 Height 객체를 사용하도록 한다.
LadderFactory의 테스트 코드를 옮기고, HeightTest로 옮긴다.
LadderFactoryTest를 삭제한다.
