+++ src/main/java/laddergame/domain/ladder/Ladder.java
@@ -1,5 +1,6 @@
 package laddergame.domain.ladder;
 
+import laddergame.domain.height.Height;
 import laddergame.domain.rung.Line;
 import laddergame.util.BooleanGenerator;
 
@@ -14,8 +15,9 @@ public class Ladder {
     }
 
     public static Ladder create(final BooleanGenerator booleanGenerator, final String height, final int participantCount) {
+        Height ladderHeight = new Height(height);
         LadderFactory ladderFactory = LadderFactory.create(booleanGenerator);
-        List<Line> ladder = ladderFactory.makeLadder(height, participantCount);
+        List<Line> ladder = ladderFactory.makeLadder(ladderHeight.getHeightNumber(), participantCount);
         return new Ladder(ladder);
     }
